app/requirements.txt
flask
flask-smorest
python-dotenv
sqlalchemy
flask-sqlalchemy
gunicorn
requests
os
pymongo